add index file with public treo. APIs
